+++ Changelog.md
@@ -6,9 +6,11 @@
 - New: More modded wooden chests can now be converted to a Password-protected Chest
 - New: Crystal Quartz Bricks, Smooth Crystal Quartz, Smooth Crystal Quartz Stairs, Smooth Crystal Quartz Slab, and their reinforced variants
 - New: Passcodes now have to be manually confirmed by using the new enter button found in the check password interface (or by using the enter key on the keyboard)
+- New: Smart Module (incorrect code = cooldown) and Harming Module (incorrect code = damage) support for blocks that require a passcode
 - API: Changed IPasswordConvertible#getOriginalBlock to IPasswordConvertible#isValidStateForConversion for finer control over what can be converted
 - API: New methods IModuleInventory#isAllowed and IModuleInventory#isDenied to check whether an entity is listed on an allowlist or denylist module respectively
 - API: Moved and renamed Owner#isOwner methods to IPasswordProtected#isOwnedBy
+- API: New methods IPasswordProtected#isOnCooldown, IPasswordProtected#startCooldown, and IPasswordProtected#getIncorrectPasscodeDamage
 - Fix: The Reinforced Water Cauldron and Reinforced Grass Block don't have a reinforced tint
 - Fix: Reinforced Stained Glass Pane items display incorrectly
 - Fix: The Reinforced Moss Carpet has a different sound compared to the vanilla Moss Carpet